seo-outline-pro/requirements.txt
google-generativeai
requests
beautifulsoup4
rich
python-dotenv
lxml
urllib3
flask
flask-cors
